Fix NPE in TryExecutor when retry policy omits optional fields (#1519)
* Fix NPE in TryExecutor when retry policy omits optional fields

The spec defines backoff, limit, and delay as optional in a retry
policy, but the runtime assumed all three were always present,
causing NullPointerException when any was omitted.

- Default to constant backoff when backoff is not specified
- Default to unlimited retries when limit is not specified
- Default to zero delay when delay is not specified


* Default to 0 retries when limit is not specified

The retry limit default was set to Integer.MAX_VALUE, but when no
limit is configured the retry should not execute. Also update the
test to verify the workflow fails when no retry limit is set.


* [Fix #1517] Review comments.


---------

+++ impl/core/src/main/java/io/serverlessworkflow/impl/TaskContext.java
@@ -37,10 +37,10 @@ public class TaskContext implements TaskContextData {
   private WorkflowModel rawOutput;
   private Instant completedAt;
   private TransitionInfo transition;
-  private short retryAttempt;
+  private int retryAttempt;
   private int iteration;
   private AuthorizationDescriptor authorization;
-  private Optional<Short> tryRetryCount = Optional.empty();
+  private Optional<Integer> tryRetryCount = Optional.empty();
 
   public TaskContext(
       WorkflowModel input,
@@ -71,7 +71,7 @@ public class TaskContext implements TaskContextData {
     this.output = output;
     this.rawOutput = rawOutput;
     this.retryAttempt =
-        parentContext.map(ctx -> ctx.tryRetryCount.orElse(ctx.retryAttempt())).orElse((short) 0);
+        parentContext.map(ctx -> ctx.tryRetryCount.orElse(ctx.retryAttempt())).orElse(0);
     this.contextVariables =
         parentContext.map(p -> new HashMap<>(p.contextVariables)).orElseGet(HashMap::new);
   }
@@ -173,19 +173,19 @@ public class TaskContext implements TaskContextData {
   }
 
   @Override
-  public short retryAttempt() {
+  public int retryAttempt() {
     return retryAttempt;
   }
 
-  public void retryAttempt(short retryAttempt) {
+  public void retryAttempt(int retryAttempt) {
     this.retryAttempt = retryAttempt;
   }
 
-  public void tryRetryCount(short tryRetryCount) {
+  public void tryRetryCount(int tryRetryCount) {
     this.tryRetryCount = Optional.of(tryRetryCount);
   }
 
-  public Optional<Short> tryRetryCount() {
+  public Optional<Integer> tryRetryCount() {
     return tryRetryCount;
   }
