+++ impl/core/src/main/java/io/serverlessworkflow/impl/executors/retry/LinearRetryIntervalFunction.java
@@ -28,7 +28,7 @@ public class LinearRetryIntervalFunction extends AbstractRetryIntervalFunction {
   }
 
   @Override
-  protected Duration calcDelay(Duration delay, short numAttempts) {
+  protected Duration calcDelay(Duration delay, int numAttempts) {
     return delay.multipliedBy(numAttempts);
   }
 }
