+++ impl/persistence/api/src/main/java/io/serverlessworkflow/impl/persistence/RetriedTaskInfo.java
@@ -15,4 +15,4 @@
  */
 package io.serverlessworkflow.impl.persistence;
 
-public record RetriedTaskInfo(short retryAttempt) implements PersistenceTaskInfo {}
+public record RetriedTaskInfo(int retryAttempt) implements PersistenceTaskInfo {}
